# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Anatomiya",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/lesson": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "All Lessons",
                "description": "Hamma Dars(Lesson)larrni ko'rish",
                "operationId": "lesson_index",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "title_uz": {
                                            "type": "string",
                                            "example": "Harley"
                                        },
                                        "title_ru": {
                                            "type": "string",
                                            "example": "Augustus"
                                        },
                                        "category_id": {
                                            "type": "number",
                                            "example": "3"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lessons"
                ],
                "summary": "new Lesson add",
                "description": "Yangi Lesson qoshish",
                "operationId": "lesson_store",
                "requestBody": {
                    "description": "lesson save",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title_uz",
                                    "title_ru",
                                    "category_id"
                                ],
                                "properties": {
                                    "title_uz": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Harley"
                                    },
                                    "title_ru": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Augustus"
                                    },
                                    "category_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "title_uz": {
                                            "type": "string",
                                            "example": "Harley"
                                        },
                                        "title_ru": {
                                            "type": "string",
                                            "example": "Augustus"
                                        },
                                        "catoegory_id": {
                                            "type": "number",
                                            "example": "3"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/lesson/{id}": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Show Lesson",
                "description": "bitta Lessonni hamma malumotlarini ko'rsatadi",
                "operationId": "lesson_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lesson id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "title_uz": {
                                            "type": "string",
                                            "example": "Harley"
                                        },
                                        "title_ru": {
                                            "type": "string",
                                            "example": "Augustus"
                                        },
                                        "catoegory_id": {
                                            "type": "number",
                                            "example": "3"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Update existing project",
                "description": "Returns updated project data",
                "operationId": "lesson_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "lesson id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "lesson update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title_uz",
                                    "title_ru",
                                    "category_id"
                                ],
                                "properties": {
                                    "title_uz": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Harley"
                                    },
                                    "title_ru": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Augustus"
                                    },
                                    "category_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "title_uz": {
                                            "type": "string",
                                            "example": "Harley"
                                        },
                                        "title_ru": {
                                            "type": "string",
                                            "example": "Augustus"
                                        },
                                        "catoegory_id": {
                                            "type": "number",
                                            "example": "3"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Delete existing project",
                "description": "Deletes a record and returns no content",
                "operationId": "lesson_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "bearer_token": {
            "type": "apiKey",
            "description": "Enter token in format (Bearer <token>)",
            "name": "Authorization",
            "in": "header"
        }
    }
}